fixed Makefile to build w/new drivers (thanks Hydra) int32 updates in sonar driver added 12mhz buzzer check Removed debug output from GPS module int32'ified althold output HSE MHz+SysClkMHz on debug[3]
git-svn-id: https://afrodevices.googlecode.com/svn/trunk/baseflight@379 7c89a4a9-59b9-e629-4cfe-3a2d53b20e61

+++ src/drv_hcsr04.c
@@ -19,7 +19,7 @@ static uint8_t exti_pin_source;
 static IRQn_Type exti_irqn;
 
 static uint32_t last_measurement;
-static volatile int16_t *distance_ptr;
+static volatile int32_t *distance_ptr;
 
 void ECHO_EXTI_IRQHandler(void)
 {
@@ -106,7 +106,7 @@ void hcsr04_init(sonar_config_t config)
 }
 
 // distance calculation is done asynchronously, using interrupt
-void hcsr04_get_distance(volatile int16_t *distance)
+void hcsr04_get_distance(volatile int32_t *distance)
 {
     uint32_t current_time = millis();
